Rename module web -> types.
