+++ deepwell/src/types/file_details.rs
@@ -1,5 +1,5 @@
 /*
- * web/file_details.rs
+ * types/file_details.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2024 Wikijump Team
