+++ deepwell/src/types/page_order.rs
@@ -1,5 +1,5 @@
 /*
- * web/page_order.rs
+ * types/page_order.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2024 Wikijump Team
