+++ deepwell/src/types/provided_value.rs
@@ -1,5 +1,5 @@
 /*
- * web/provided_value.rs
+ * types/provided_value.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2024 Wikijump Team
