+++ deepwell/src/types/reference.rs
@@ -1,5 +1,5 @@
 /*
- * web/reference/id.rs
+ * types/reference/id.rs
  *
  * DEEPWELL - Wikijump API provider and database manager
  * Copyright (C) 2019-2024 Wikijump Team
